Added code that SHOULD move cursor, indicating the issue with this and clock are both due to port functions

+++ include/Ports.hpp
@@ -6,8 +6,8 @@
 namespace angl {
     namespace io {
         namespace port {
-            void write(uint8_t port, uint8_t data);
-            uint8_t read(uint8_t port);
+            void write(uint32_t port, uint8_t data);
+            uint8_t read(uint32_t port);
         }
     }
 }
